#Entities:
Updated properties and methods in several classes

Updated several classes in the project. The nullable string properties in the `Category`, `Role`, `Tag`, and `User` classes have been changed to non-nullable and are now initialized to an empty string by default. A new property `Posts` of type `IList<Post>` has been added to the `Category` class. The `Id` property in the `Role` class has been simplified and the `[Key]` attribute has been removed from it. The `ToString` method in the `Role` class has been modified to use `this` keyword when referencing the properties.