#Mapping:
Created CategoryMap class for Category entity configuration

Created a new class `CategoryMap` in the `Blog.Data.Mappings` namespace to configure the `Category` entity in the Entity Framework Core. The `Configure` method sets up the properties and relationships of the `Category` entity. The `Category` table in the database is specified, and the primary key is mapped to the `Id` property. The `Id` property is set to auto-increment. The `Name` and `Slug` properties are mapped to columns in the database, set as required, their type is set to `NVARCHAR`, and their length is limited to 80 characters. A unique index is created for the `Slug` property.